pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "km-digitizer"
version = "0.1.0"
description = "Automated Kaplan-Meier curve digitization + synthetic benchmark"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "opencv-python>=4.8",
  "matplotlib>=3.8",
  "pillow>=10.0",
  "pandas>=2.0",
  "scipy>=1.10",
  "pytesseract>=0.3.10",
]

[tool.setuptools.packages.find]
where = ["src"]
